dictionary/src/test/java/com/danielvaughan/wordgametools/dictionary/DictionaryControllerTest.java
package com.danielvaughan.wordgametools.dictionary;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.is;
import static org.mockito.BDDMockito.given;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(DictionaryController.class)
public class DictionaryControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private DictionaryService mockDictionaryService;

    @Test
    public void given_empty_string_then_return_false() throws Exception {
        given(mockDictionaryService.lookup("")).willReturn(false);

        mockMvc.perform(get("/lookup?word=")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", is(false)));

    }

    @Test
    public void given_cat_then_return_true() throws Exception {
        given(mockDictionaryService.lookup("cat")).willReturn(true);

        mockMvc.perform(get("/lookup?word=cat")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", is(true)));

    }
}
